Initial and final commit
This is homework for CSC 4585 that demonstrates some practical multithreading, rather than just theory, although it's a practical demonstration of theory.  The locks here are useful because they use a single memory location for N threads, and perform adequately when contention isn't too high.

Anyone is free to use these locks for any reason, except for purposes of cheating or other unethical behavior.